requirements.txt
ipython
ipdb
Pillow
